reactRouterDom add to directory-item now the route uses the route key property in the Items Object pased to the directoryComponent so we can click on the items to acces or naveigate the pages by clicking on the items
