+++ src/components/directory-item/directory-item.component.jsx
@@ -1,3 +1,5 @@
+import { useNavigate } from "react-router-dom";
+
 import {
   DirectoryItemContainer,
   BackgroundImage,
@@ -5,9 +7,12 @@ import {
 } from "./directory-item.styles";
 
 const DirectoryItem = ({ category }) => {
-  const { imageUrl, title } = category;
+  const { imageUrl, title, route } = category;
+  const navigate = useNavigate();
+
+  const onNavigateHandler = () => navigate(route);
   return (
-    <DirectoryItemContainer>
+    <DirectoryItemContainer onClick={onNavigateHandler}>
       <BackgroundImage imageUrl={imageUrl} />
       <Body>
         <h2>{title}</h2>
