+++ src/context/categories.context.jsx
@@ -21,7 +21,7 @@ export const CategoriesProvider = ({ children }) => {
   useEffect(() => {
     const getCategoriesMap = async () => {
       const categoryMap = await getCategoriesAndDocuments("categories");
-      console.log(categoryMap);
+      /* console.log(categoryMap); */
       setCategoriesMap(categoryMap);
     };
     getCategoriesMap();
